Pull out RichMessageExtractor
No need to have it embedded
  when it should be unit tested!
