+++ app/services/todo_from_rich_message.rb
@@ -1,34 +1,5 @@
 class TodoFromRichMessage
 
-  class RichMessageExtractor
-
-    RICH_MESSAGE_FIELDS_REGEX = /([^>@]*)@?([^>]*)>?(.*)/
-
-    def initialize(message)
-      @message = message
-    end
-
-    def description
-      fields[1].strip
-    end
-
-    def context
-      fields[2].strip
-    end
-
-    def project
-      stripped = fields[3].strip
-      stripped.blank? ? nil : stripped
-    end
-
-    private
-
-      def fields
-        @message.match(RICH_MESSAGE_FIELDS_REGEX)
-      end
-
-  end
-
   attr_reader :user, :default_context_id, :description, :notes
 
   def initialize(user, default_context_id, description, notes)
